Refactor TripDetails Component Update
